==> Initial Code/Research Question One/9. Grid_Search_Results.ipynb <==
logRegScores1 = pd.read_pickle("logRegScores1.pickle")
f2_scoreNB1 = pd.read_pickle("f2_scoreNB1.pickle")
rfScores1 = pd.read_pickle("rfScores1.pickle")
bstScores1 = pd.read_pickle("bstScores1.pickle")

logRegScores2 = pd.read_pickle("logRegScores2.pickle")
f2_scoreNB2 = pd.read_pickle("f2_scoreNB2.pickle")
rfScores2 = pd.read_pickle("rfScores2.pickle")
bstScores2 = pd.read_pickle("bstScores2.pickle")

logRegScores3 = pd.read_pickle("logRegScores3.pickle")
f2_scoreNB3 = pd.read_pickle("f2_scoreNB3.pickle")
rfScores3 = pd.read_pickle("rfScores3.pickle")
bstScores3 = pd.read_pickle("bstScores3.pickle")

logRegScores4 = pd.read_pickle("logRegScores4.pickle")
f2_scoreNB4 = pd.read_pickle("f2_scoreNB4.pickle")
rfScores4 = pd.read_pickle("rfScores4.pickle")
bstScores4 = pd.read_pickle("bstScores4.pickle")

logRegScores5 = pd.read_pickle("logRegScores5.pickle")
f2_scoreNB5 = pd.read_pickle("f2_scoreNB5.pickle")
rfScores5 = pd.read_pickle("rfScores5.pickle")
bstScores5 = pd.read_pickle("bstScores5.pickle")

Logistic Regression

logRegScores1 = logRegScores1.drop(['Unnamed: 0'], axis=1)
logRegScores2 = logRegScores2.drop(['Unnamed: 0'], axis=1)
logRegScores3 = logRegScores3.drop(['Unnamed: 0'], axis=1)
logRegScores4 = logRegScores4.drop(['Unnamed: 0'], axis=1)
logRegScores5 = logRegScores5.drop(['Unnamed: 0'], axis=1)

merged_logRegA = logRegScores1.merge(logRegScores2,on=['Solver'], suffixes=('_x1', '_x2'))
merged_logRegB = merged_logRegA.merge(logRegScores3,on=['Solver'])
merged_logRegC = merged_logRegB.merge(logRegScores4,on=['Solver'], suffixes=('_x3', '_x4'))
merged_logReg = merged_logRegC.merge(logRegScores5,on=['Solver'])
merged_logReg.rename(columns={'F2_Score': 'F2_Score_x5'}, inplace=True)

#Computing median F-2 Score across all five sliding windows for each combination
#of hyperparameters
for i in range(len(merged_logReg['F2_Score_x1'])):
      merged_logReg.loc[merged_logReg.index[i], 'Median_Score'] = statistics.median([merged_logReg.loc[merged_logReg.index[i], 'F2_Score_x1'],merged_logReg.loc[merged_logReg.index[i], 'F2_Score_x2'], merged_logReg.loc[merged_logReg.index[i], 'F2_Score_x3'], merged_logReg.loc[merged_logReg.index[i], 'F2_Score_x4'], merged_logReg.loc[merged_logReg.index[i], 'F2_Score_x5']] )

merged_logReg.sort_values("Median_Score", ascending=False, inplace=True)
merged_logReg

#The optimal hyperparameter is Solver=liblinear

Naive Bayes

NB_score_list = [f2_scoreNB1, f2_scoreNB2, f2_scoreNB3, f2_scoreNB4, f2_scoreNB5]
NB_score_list.sort(ascending=False)
#Naive Bayes does not have any hyperparameters to optimize
#(The F2-Scores were just investigated out of interest)

Random Forest

rfScores1 = rfScores1.drop(['Unnamed: 0'], axis=1)
rfScores2 = rfScores2.drop(['Unnamed: 0'], axis=1)
rfScores3 = rfScores3.drop(['Unnamed: 0'], axis=1)
rfScores4 = rfScores4.drop(['Unnamed: 0'], axis=1)
rfScores5 = rfScores5.drop(['Unnamed: 0'], axis=1)

merged_rfA = rfScores1.merge(rfScores2,on=['n_Estimators', 'Max_depth', 'Criterion'], suffixes=('_x1', '_x2'))
merged_rfB = merged_rfA.merge(rfScores3,on=['n_Estimators', 'Max_depth', 'Criterion'])
merged_rfC = merged_rfB.merge(rfScores4,on=['n_Estimators', 'Max_depth', 'Criterion'], suffixes=('_x3', '_x4'))
merged_rf = merged_rfC.merge(rfScores5,on=['n_Estimators', 'Max_depth', 'Criterion'])
merged_rf.rename(columns={'F2_Score': 'F2_Score_x5'}, inplace=True)

#Computing median F-2 Score across all five sliding windows for each combination
#of hyperparameters
for i in range(len(merged_rf['F2_Score_x1'])):
      merged_rf.loc[merged_rf.index[i], 'Median_Score'] = statistics.median([merged_rf.loc[merged_rf.index[i], 'F2_Score_x1'],merged_rf.loc[merged_rf.index[i], 'F2_Score_x2'], merged_rf.loc[merged_rf.index[i], 'F2_Score_x3'], merged_rf.loc[merged_rf.index[i], 'F2_Score_x4'], merged_rf.loc[merged_rf.index[i], 'F2_Score_x5']] )

merged_rf.sort_values("Median_Score", ascending=False, inplace=True)
merged_rf

#The optimal hyperparameters are n_Estimators=100, max_depth=1, and criterion="log_loss"

 Gradient-Boosted Trees

bstScores1 = bstScores1.drop(['Unnamed: 0'], axis=1)
bstScores2 = bstScores2.drop(['Unnamed: 0'], axis=1)
bstScores3 = bstScores3.drop(['Unnamed: 0'], axis=1)
bstScores4 = bstScores4.drop(['Unnamed: 0'], axis=1)
bstScores5 = bstScores5.drop(['Unnamed: 0'], axis=1)

merged_bstA = bstScores1.merge(bstScores2,on=['n_Estimators', 'Max_depth', 'Subsample'], suffixes=('_x1', '_x2'))
merged_bstB = merged_bstA.merge(bstScores3,on=['n_Estimators', 'Max_depth', 'Subsample'])
merged_bstC = merged_bstB.merge(bstScores4,on=['n_Estimators', 'Max_depth', 'Subsample'], suffixes=('_x3', '_x4'))
merged_bst = merged_bstC.merge(bstScores5,on=['n_Estimators', 'Max_depth', 'Subsample'])
merged_bst.rename(columns={'F2_Score': 'F2_Score_x5'}, inplace=True)

#Computing median F-2 Score across all five sliding windows for each combination
#of hyperparameters
for i in range(len(merged_bst['F2_Score_x1'])):
      merged_bst.loc[merged_bst.index[i], 'Median_Score'] = statistics.median([merged_bst.loc[merged_bst.index[i], 'F2_Score_x1'],merged_bst.loc[merged_bst.index[i], 'F2_Score_x2'], merged_bst.loc[merged_bst.index[i], 'F2_Score_x3'], merged_bst.loc[merged_bst.index[i], 'F2_Score_x4'], merged_bst.loc[merged_bst.index[i], 'F2_Score_x5']] )

merged_bst.sort_values("Median_Score", ascending=False, inplace=True)
merged_bst

#The optimal hyperparameters are n_Estimators=25, max_depth=1, and subsample=0.2
